All dialogs was moved to dialog folder
