--- src/app/add-attack-dialog/add-attack-dialog.component.css
@@ -1,1 +1,0 @@
-@import '../shared/css/addDialog.css';
