--- src/app/add-equipment-dialog/add-equipment-dialog.component.css
@@ -1,1 +1,0 @@
-@import '../shared/css/addDialog.css';
