--- src/app/add-proficiency-dialog/add-proficiency-dialog.component.css
@@ -1,1 +1,0 @@
-@import '../shared/css/addDialog.css';
