--- src/app/add-string-dialog/add-string-dialog.component.css
@@ -1,1 +1,0 @@
-@import '../shared/css/addDialog.css';
