+++ src/app/dialogs/add-attack-dialog/add-attack-dialog.component.css
@@ -1,0 +1,1 @@
+@import '../../shared/css/addDialog.css';
