+++ src/app/dialogs/add-attack-dialog/add-attack-dialog.component.html
