+++ src/app/dialogs/add-equipment-dialog/add-equipment-dialog.component.css
@@ -1,0 +1,1 @@
+@import '../../shared/css/addDialog.css';
