+++ src/app/dialogs/add-equipment-dialog/add-equipment-dialog.component.html
