+++ src/app/dialogs/add-proficiency-dialog/add-proficiency-dialog.component.css
@@ -1,0 +1,1 @@
+@import '../../shared/css/addDialog.css';
