+++ src/app/dialogs/add-string-dialog/add-string-dialog.component.css
@@ -1,0 +1,1 @@
+@import '../../shared/css/addDialog.css';
