+++ src/app/dialogs/add-string-dialog/add-string-dialog.component.spec.ts
