+++ src/app/dialogs/remove-character-dialog/remove-character-dialog.component.css
@@ -1,0 +1,1 @@
+@import '../../shared/css/dialog.css';
