+++ src/app/dialogs/remove-character-dialog/remove-character-dialog.component.spec.ts
