--- src/app/remove-character-dialog/remove-character-dialog.component.css
@@ -1,1 +1,0 @@
-@import '../shared/css/dialog.css';
